+++ packages/x-adapter-platform/src/mappers/requests/index.ts
@@ -1,6 +1,4 @@
 export * from './base-request.mapper';
-export * from './popular-searches-request.mapper';
-export * from './query-suggestions-request.mapper';
+export * from './identifier-results-request.mapper';
 export * from './search-request.mapper';
 export * from './tagging-request.mapper';
-export * from './next-queries-request.mapper';