+++ packages/x-adapter-platform/src/mappers/responses/search-response.mapper.ts
@@ -1,7 +1,7 @@
 import { createMutableSchema, Schema, schemaMapperFactory } from '@empathyco/x-adapter-next';
 import { getTaggingInfoFromUrl } from '@empathyco/x-utils';
 import { PlatformSearchResponse, SearchResponse } from '../../types/response.types';
-import { resultMutableSchema } from '../../schemas/result.schema';
+import { resultSchema } from '../../schemas/models/result.schema';
 import { facetMutableSchema } from '../../schemas/facet.schema';
 import { promotedMutableSchema } from '../../schemas/promoted.schema';
 import { bannerMutableSchema } from '../../schemas/banner.schema';
@@ -10,7 +10,7 @@ import { redirectionMutableSchema } from '../../schemas/redirection.schema';
 export const searchResponseSchema: Schema<PlatformSearchResponse, SearchResponse> = {
   results: {
     $path: 'catalog.content',
-    $subSchema: resultMutableSchema
+    $subSchema: resultSchema
   },
   facets: {
     $path: 'catalog.facets',