+++ packages/x-adapter-platform/src/platform.adapter.ts
@@ -5,7 +5,8 @@ import { popularSearchesEndpointAdapter } from './endpoint-adapters/popular-sear
 import { topClickedEndpointAdapter } from './endpoint-adapters/top-clicked.endpoint-adapter';
 import { nextQueriesEndpointAdapter } from './endpoint-adapters/next-queries.endpoint-adapter';
 import { relatedTagsEndpointAdapter } from './endpoint-adapters/related-tags.endpoint-adapter';
-import { skuSearchEndpointAdapter } from './endpoint-adapters/sku-search.endpoint-adapter';
+// eslint-disable-next-line max-len
+import { identifierResultsEndpointAdapter } from './endpoint-adapters/identifier-results.endpoint-adapter';
 import { taggingEndpointAdapter } from './endpoint-adapters/tagging.endpoint-adapter';
 // eslint-disable-next-line max-len
 import { querySuggestionsEndpointAdapter } from './endpoint-adapters/query-suggestions.endpoint-adapter';
@@ -17,6 +18,6 @@ export const platformAdapter: PlatformAdapter = {
   nextQueries: nextQueriesEndpointAdapter,
   querySuggestions: querySuggestionsEndpointAdapter,
   relatedTags: relatedTagsEndpointAdapter,
-  skuSearch: skuSearchEndpointAdapter,
+  identifierResults: identifierResultsEndpointAdapter,
   tagging: taggingEndpointAdapter
 };