+++ packages/x-adapter-platform/src/schemas/index.ts
@@ -2,7 +2,6 @@ export * from './filters';
 export * from './models';
 export * from './requests';
 export * from './responses';
-export * from './result.schema';
 export * from './banner.schema';
 export * from './facet.schema';
 export * from './promoted.schema';