+++ packages/x-adapter-platform/src/schemas/models/index.ts
@@ -1,2 +1,3 @@
 export * from './suggestion.schema';
 export * from './next-query.schema';
+export * from './result.schema';