+++ packages/x-adapter-platform/src/types/requests/index.ts
@@ -1,2 +1,3 @@
 export * from './popular-searches-request.model';
 export * from './query-suggestions-request.model';
+export * from './identifier-results-request.model';