+++ packages/x-adapter-platform/src/types/requests/request.types.ts
@@ -27,3 +27,30 @@ export interface PlatformPageableRequest {
 export interface PlatformQueryableRequest {
   query: string;
 }
+
+/**
+ * Interface for any `request` with `sorting`.
+ *
+ * @public
+ */
+export declare interface PlatformSortableRequest {
+  sort?: string;
+}
+
+/**
+ * Interface for any `request` with `tagging`.
+ *
+ * @public
+ */
+export interface PlatformTrackableRequest {
+  origin?: string;
+}
+
+/**
+ * Interface for any `request` with `filters`.
+ *
+ * @public
+ */
+export interface PlatformFilterableRequest {
+  filter?: string[];
+}